Move horizontal scroll view into own file
